+++ slack-app/lib/content/loader.ts
@@ -1,0 +1,55 @@
+/** Re-exports declarative Slack UI content loaders. */
+export {
+  buildCreateTesEventModalView,
+  buildOnboardingModalView,
+  CREATE_TES_EVENT_MODAL_BLOCKS,
+  getModalBlockIds,
+  ONBOARDING_MODAL_BLOCKS,
+  parseModalJson,
+  resetModalCacheForTests,
+} from "./modal-compiler.ts";
+export type { ModalDefinition, OnboardingModalParams } from "./modal-compiler.ts";
+
+export {
+  dashboardTemplate,
+  infrastructureTemplate,
+  renderDashboardCanvas,
+  renderInfrastructureCanvas,
+  renderRequirementsCanvas,
+  requirementsTemplate,
+  resetCanvasCacheForTests,
+} from "./canvas-renderer.ts";
+
+export {
+  getDeliverablesStatusOptions,
+  getListColumns,
+  getListName,
+  getSlackListSchema,
+  DELIVERABLES_COLUMNS,
+  INCIDENTS_COLUMNS,
+  parseListJson,
+  resetListCacheForTests,
+} from "./list-compiler.ts";
+export type {
+  ListColumnDefinition,
+  ListColumnOption,
+  ListDefinition,
+  SlackListColumn,
+} from "./list-compiler.ts";
+
+export {
+  pinnedIndexBlocks,
+  pinnedIndexMessage,
+  renderPinnedIndexBlocks,
+  renderPinnedIndexMessage,
+  resetMessageCacheForTests,
+} from "./message-renderer.ts";
+
+export {
+  deriveComponents,
+  getDeliverableStatusChoices,
+  getSupportedSuites,
+  parseDeliverableStatusesJson,
+  parseSailpointSuitesJson,
+  resetDomainCacheForTests,
+} from "./domain.ts";